+++ pkg/lint/findings/findings.go
@@ -1,11 +1,12 @@
 package findings
 
+import "github.com/giantswarm/schemalint/pkg/cli"
+
 type Severity int
 
 const (
 	SeverityError Severity = iota
-	SeverityWarning
-	SeverityInfo
+	SeverityRecomendation
 )
 
 type LintFindings struct {
@@ -14,6 +15,12 @@ type LintFindings struct {
 }
 
 func (f LintFindings) String() string {
+	switch f.Severity {
+	case SeverityError:
+		return cli.SprintErrorMessage(f.Message)
+	case SeverityRecomendation:
+		return f.Message
+	}
 	return f.Message
 }