+++ pkg/lint/rulesets/rulesets.go
@@ -5,12 +5,12 @@ import (
 
 	"github.com/santhosh-tekuri/jsonschema/v5"
 
-	"github.com/giantswarm/schemalint/pkg/lint/findings"
-	"github.com/giantswarm/schemalint/pkg/lint/rulesets/rules"
+	"github.com/giantswarm/schemalint/pkg/lint/rulesmeta"
+	"github.com/giantswarm/schemalint/pkg/lint/rulesmeta/rules"
 )
 
 type RuleSet struct {
-	rules []rules.Rule
+	rules []rulesmeta.Rule
 }
 type RuleSetName string
 
@@ -19,8 +19,9 @@ const (
 )
 
 var ClusterApp = &RuleSet{
-	rules: []rules.Rule{
+	rules: []rulesmeta.Rule{
 		rules.TitleExists{},
+		rules.DescriptionExists{},
 	},
 }
 
@@ -66,30 +67,22 @@ func GetRuleSet(name RuleSetName) *RuleSet {
 	return ruleSet
 }
 
-func VerifyRuleSet(name string, schema *jsonschema.Schema) []findings.LintFindings {
+func VerifyRuleSet(name string, schema *jsonschema.Schema) (errors []string, recommendations []string) {
 	nameEnum := RuleSetName(name)
 	ruleSet := GetRuleSet(nameEnum)
 
-	lintFindings := []findings.LintFindings{}
+	errors = []string{}
+	recommendations = []string{}
 	for _, rule := range ruleSet.rules {
 		violations := rule.Verify(schema)
-		lintFindings = append(
-			lintFindings,
-			tranformRuleViolationsToFindings(violations, rule.GetSeverity())...,
-		)
+		severity := rule.GetSeverity()
+		if severity == rulesmeta.SeverityError {
+			errors = append(errors, violations...)
+		}
+		if severity == rulesmeta.SeverityRecomendation {
+			recommendations = append(recommendations, violations...)
+		}
 	}
 
-	return lintFindings
-}
-
-func tranformRuleViolationsToFindings(violations []rules.RuleViolation, severity findings.Severity) []findings.LintFindings {
-	lintFindings := []findings.LintFindings{}
-	for _, violation := range violations {
-		lintFindings = append(lintFindings, findings.LintFindings{
-			Severity: severity,
-			Message:  violation.Reason,
-		})
-	}
-
-	return lintFindings
+	return errors, recommendations
 }