+++ pkg/lint/rulesmeta/rules.go
@@ -1,0 +1,17 @@
+package rulesmeta
+
+import (
+	"github.com/santhosh-tekuri/jsonschema/v5"
+)
+
+type Severity int
+
+const (
+	SeverityError Severity = iota
+	SeverityRecomendation
+)
+
+type Rule interface {
+	Verify(*jsonschema.Schema) []string
+	GetSeverity() Severity
+}