+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add check that an description exists on every property to `cluster-app` rule set.
+
 ## [0.4.0] - 2023-01-16
 
 - Add possible values of `rule-set` to `--help` message